refactor: extract dialog helpers from dom.js into dom-dialogs.js
Move createDialogBase, createCustomModal, showPromptDialog, and
showConfirmDialog into a dedicated dom-dialogs.js module to bring
dom.js below the 300-line threshold (317 -> 194 lines).
Re-exports preserve backward compatibility for all existing imports.

Closes #83
