+++ src/utils/dom-dialogs.js
@@ -1,0 +1,140 @@
+/**
+ * Dialog and prompt helpers — extracted from dom.js to reduce file length.
+ *
+ * Re-uses the low-level DOM factories from dom.js (_el, createActionButton,
+ * createModalOverlay, setupKeyboardShortcuts).
+ */
+
+import { _el, createActionButton, createModalOverlay, setupKeyboardShortcuts } from './dom.js';
+
+// ── Private dialog lifecycle ──
+
+/**
+ * Private dialog lifecycle helper.
+ * Creates overlay + modal via createModalOverlay, calls builder to populate
+ * content, appends to document.body, and wraps everything in a Promise.
+ *
+ * @param {{ overlayClass: string, modalClass: string, cancelValue?: unknown, onCancel?: () => void, builder: (ctx: { overlay: HTMLElement, modal: HTMLElement, cleanup: (value: unknown) => void, cancel: () => void }) => (() => void) | void }} opts
+ * @param {string} opts.overlayClass - CSS class for the overlay element
+ * @param {string} opts.modalClass - CSS class for the modal element
+ * @param {unknown} [opts.cancelValue=null] - value passed to resolve on cancel / click-outside
+ * @param {() => void} [opts.onCancel] - optional callback fired after cancel cleanup
+ * @param {(ctx: { overlay: HTMLElement, modal: HTMLElement, cleanup: (value: unknown) => void, cancel: () => void }) => (() => void) | void} opts.builder - receives ({ overlay, modal, cleanup, cancel }).
+ *   cleanup(value) removes the overlay and resolves the promise.
+ *   cancel() is a shorthand for cleanup(cancelValue) + onCancel?.().
+ *   May return a function that runs after the overlay is appended to the DOM
+ *   (useful for focusing elements).
+ * @returns {Promise<unknown>}
+ */
+function createDialogBase({ overlayClass, modalClass, cancelValue = null, onCancel, builder }) {
+  return new Promise((resolve) => {
+    let overlay;
+    const cleanup = (value) => { overlay.remove(); resolve(value); };
+    const cancel = () => { cleanup(cancelValue); onCancel?.(); };
+    ({ overlay } = createModalOverlay(overlayClass, modalClass, cancel));
+    const modal = overlay.firstChild;
+    const afterMount = builder({ overlay, modal, cleanup, cancel });
+    document.body.appendChild(overlay);
+    if (typeof afterMount === 'function') afterMount();
+  });
+}
+
+// ── Exported dialog builders ──
+
+/**
+ * High-level modal builder: creates overlay + modal with a title bar,
+ * content area, and optional close button. Appends to document.body and
+ * returns a Promise that resolves when the modal is closed.
+ *
+ * @param {{ title?: string, content?: Node|Node[], onClose?: () => void,
+ *           overlayClass?: string, modalClass?: string }} opts
+ * @returns {Promise<null>}
+ */
+export function createCustomModal({ title, content, onClose, overlayClass = 'modal-overlay', modalClass = 'modal' } = {}) {
+  return createDialogBase({
+    overlayClass,
+    modalClass,
+    onCancel: onClose,
+    builder({ modal, cancel }) {
+      if (title) {
+        const header = _el('div', `${modalClass}-header`,
+          _el('span', `${modalClass}-title`, title),
+          createActionButton({ text: '\u00D7', cls: `${modalClass}-close-btn`, onClick: cancel }),
+        );
+        modal.appendChild(header);
+      }
+
+      const body = _el('div', `${modalClass}-body`);
+      if (content) {
+        const nodes = Array.isArray(content) ? content : [content];
+        for (const node of nodes) {
+          if (node) body.appendChild(node);
+        }
+      }
+      modal.appendChild(body);
+
+      setupKeyboardShortcuts(modal, { onEscape: cancel });
+    },
+  });
+}
+
+/**
+ * Show a prompt dialog for a single text value.
+ * @returns {Promise<string|null>} trimmed value or null if cancelled
+ */
+export function showPromptDialog({ title, placeholder = '', defaultValue = '', confirmLabel = 'Create', cancelLabel = 'Cancel' }) {
+  return createDialogBase({
+    overlayClass: 'prompt-dialog-overlay',
+    modalClass: 'prompt-dialog-box',
+    builder({ modal, cleanup, cancel }) {
+      const confirm = () => { const v = input.value.trim(); cleanup(v || null); };
+      const input = _el('input', { className: 'prompt-dialog-input', type: 'text', value: defaultValue, placeholder });
+      setupKeyboardShortcuts(input, {
+        onEnter: () => confirm(),
+        onEscape: cancel,
+      });
+      modal.append(
+        _el('label', 'prompt-dialog-label', title),
+        input,
+        _el('div', 'prompt-dialog-btns',
+          createActionButton({ text: cancelLabel, cls: 'prompt-dialog-cancel', onClick: cancel }),
+          createActionButton({ text: confirmLabel, cls: 'prompt-dialog-confirm', onClick: confirm }),
+        ),
+      );
+      return () => {
+        input.focus();
+        if (defaultValue) input.select();
+      };
+    },
+  });
+}
+
+/**
+ * Show a confirm dialog.
+ * @param {Node|string} message - text string or DOM node
+ * @returns {Promise<boolean>}
+ */
+export function showConfirmDialog(message, { confirmLabel = 'OK', cancelLabel = 'Cancel' } = {}) {
+  return createDialogBase({
+    overlayClass: 'confirm-overlay',
+    modalClass: 'confirm-box',
+    cancelValue: false,
+    builder({ overlay, modal, cleanup, cancel }) {
+      if (typeof message === 'string') modal.appendChild(_el('p', null, message));
+      else modal.appendChild(message);
+
+      const btnRow = _el('div', 'confirm-buttons',
+        createActionButton({ text: cancelLabel, cls: 'confirm-cancel', onClick: cancel }),
+        createActionButton({ text: confirmLabel, cls: 'confirm-ok', onClick: () => cleanup(true) }),
+      );
+      modal.appendChild(btnRow);
+
+      setupKeyboardShortcuts(overlay, {
+        onEscape: cancel,
+        onEnter: () => cleanup(true),
+      });
+      overlay.setAttribute('tabindex', '-1');
+      return () => btnRow.querySelector('.confirm-ok').focus();
+    },
+  });
+}
